- Fixed Reprint Label feature
