+++ RockWeb/Plugins/com_bemadev/Checkin/ReprintLabel.ascx.cs
@@ -97,197 +97,18 @@ namespace RockWeb.Plugins.com_bemaservices.CheckIn
         /// <param name="e">The <see cref="EventArgs"/> instance containing the event data.</param>
         protected void btnPrint_Click( object sender, EventArgs e )
         {
-            var commonMergeFields = Rock.Lava.LavaHelper.GetCommonMergeFields( null );
-
             RockContext rockContext = new RockContext();
-
-            var groupMemberService = new GroupMemberService( rockContext );
-            DeviceService deviceService = new DeviceService( rockContext );
-
-            CheckInPerson checkinPerson = new CheckInPerson();
             if ( ppPerson.PersonId.HasValue )
             {
                 PersonService personService = new PersonService( rockContext );
                 var person = personService.Get( ppPerson.PersonId.Value );
                 if ( person != null )
                 {
-                    checkinPerson.Person = person;
-                    var attendance = new AttendanceService( rockContext ).Queryable().Where( a => a.PersonAlias.PersonId == person.Id ).OrderByDescending( a => a.CreatedDateTime ).FirstOrDefault();
-                    if ( attendance != null )
-                    {
-                        checkinPerson.SecurityCode = attendance.AttendanceCode.Code;
-                        try
-                        {
-                            var PrinterIPs = new Dictionary<int, string>();
-
-                            string detailMsg = GetAttributeValue( "DetailMessage" );
-
-                            var personLabelsAdded = new List<Guid>();
-                            var printFromClient = new List<CheckInLabel>();
-                            var printFromServer = new List<CheckInLabel>();
-                            var checkinLabels = new List<CheckInLabel>();
-                            var methodHandler = new Rock.Workflow.Action.CheckIn.CreateLabels();
-
-                            var mergeObjects = new Dictionary<string, object>();
-                            foreach ( var keyValue in commonMergeFields )
-                            {
-                                mergeObjects.Add( keyValue.Key, keyValue.Value );
-                            }
-
-                            var checkinSchedule = new CheckInSchedule();
-                            checkinSchedule.Schedule = attendance.Occurrence.Schedule.Clone( false );
-                            checkinSchedule.Selected = true;
-
-                            var checkinLocation = new CheckInLocation();
-                            checkinLocation.Location = attendance.Occurrence.Location.Clone( false );
-                            checkinLocation.Location.CopyAttributesFrom( attendance.Occurrence.Location );
-                            checkinLocation.Schedules.Add( checkinSchedule );
-                            checkinLocation.Selected = true;
-
-                            var checkinGroup = new CheckInGroup();
-                            checkinGroup.Group = attendance.Occurrence.Group.Clone( false );
-                            checkinGroup.Group.CopyAttributesFrom( attendance.Occurrence.Group );
-                            checkinGroup.Locations.Add( checkinLocation );
-                            checkinGroup.Selected = true;
-
-                            var checkinGroupType = new CheckInGroupType();
-                            checkinGroupType.GroupType = GroupTypeCache.Get( attendance.Occurrence.Group.GroupType );
-                            checkinGroupType.Groups.Add( checkinGroup );
-                            checkinGroupType.Selected = true;
-
-                            var checkinPeople = new List<CheckInPerson>();
-                            checkinPeople.Add( checkinPerson );
-
-                            mergeObjects.Add( "Location", checkinLocation );
-                            mergeObjects.Add( "Group", checkinGroup );
-                            mergeObjects.Add( "Person", checkinPerson );
-                            mergeObjects.Add( "People", checkinPeople );
-                            mergeObjects.Add( "GroupType", checkinGroupType );
-
-                            var groupMembers = groupMemberService.Queryable().AsNoTracking()
-                                .Where( m =>
-                                    m.PersonId == person.Id &&
-                                    m.GroupId == attendance.Occurrence.Group.Id )
-                                .ToList();
-                            mergeObjects.Add( "GroupMembers", groupMembers );
-
-                            var personLabels = methodHandler.GetLabels( checkinPerson.Person, new List<KioskLabel>() );
-                            var groupTypeLabels = methodHandler.GetLabels( attendance.Occurrence.Group.GroupType, personLabels );
-                            var groupLabels = methodHandler.GetLabels( attendance.Occurrence.Group, groupTypeLabels );
-                            var locationLabels = methodHandler.GetLabels( attendance.Occurrence.Location, groupLabels );
-
-                            Device kioskDevice = null;
-                            if ( CurrentKioskId.HasValue )
-                            {
-                                kioskDevice = new DeviceService( rockContext ).Get( CurrentKioskId.Value );
-                            }
-                            else
-                            {
-                                kioskDevice = attendance.Device;
-                            }
-
-                            if ( kioskDevice != null )
-                            {
-                                foreach ( var labelCache in locationLabels.OrderBy( l => l.LabelType ).ThenBy( l => l.Order ) )
-                                {
-                                    checkinPerson.SetOptions( labelCache );
-
-                                    if ( labelCache.LabelType == KioskLabelType.Person )
-                                    {
-                                        if ( personLabelsAdded.Contains( labelCache.Guid ) )
-                                        {
-                                            continue;
-                                        }
-                                        else
-                                        {
-                                            personLabelsAdded.Add( labelCache.Guid );
-                                        }
-                                    }
-
-
-                                    var label = new CheckInLabel( labelCache, mergeObjects, person.Id );
-                                    label.FileGuid = labelCache.Guid;
-
-                                    label.PrintFrom = kioskDevice.PrintFrom;
-                                    label.PrintTo = kioskDevice.PrintToOverride;
-
-                                    if ( label.PrintTo == PrintTo.Default )
-                                    {
-                                        label.PrintTo = attendance.Occurrence.Group.GroupType.AttendancePrintTo;
-                                    }
-
-                                    if ( label.PrintTo == PrintTo.Kiosk )
-                                    {
-                                        var device = kioskDevice;
-                                        if ( device != null )
-                                        {
-                                            label.PrinterDeviceId = device.PrinterDeviceId;
-                                        }
-                                    }
-                                    else if ( label.PrintTo == PrintTo.Location )
-                                    {
-                                        var deviceId = attendance.Occurrence.Location.PrinterDeviceId;
-                                        if ( deviceId != null )
-                                        {
-                                            label.PrinterDeviceId = deviceId;
-                                        }
-                                    }
-
-                                    if ( label.PrinterDeviceId.HasValue )
-                                    {
-                                        if ( PrinterIPs.ContainsKey( label.PrinterDeviceId.Value ) )
-                                        {
-                                            label.PrinterAddress = PrinterIPs[label.PrinterDeviceId.Value];
-                                        }
-                                        else
-                                        {
-                                            var printerDevice = new DeviceService( rockContext ).Get( label.PrinterDeviceId.Value );
-                                            if ( printerDevice != null )
-                                            {
-                                                PrinterIPs.Add( printerDevice.Id, printerDevice.IPAddress );
-                                                label.PrinterAddress = printerDevice.IPAddress;
-                                            }
-                                        }
-                                    }
-
-                                    checkinLabels.Add( label );
-                                }
-                            }
-
-
-                            // Print the labels                   
-                            if ( checkinLabels != null && checkinLabels.Any() )
-                            {
-                                printFromClient.AddRange( checkinLabels.Where( l => l.PrintFrom == Rock.Model.PrintFrom.Client ) );
-                                printFromServer.AddRange( checkinLabels.Where( l => l.PrintFrom == Rock.Model.PrintFrom.Server ) );
-                            }
-
-                            if ( printFromClient.Any() )
-                            {
-                                var urlRoot = string.Format( "{0}://{1}", Request.Url.Scheme, Request.Url.Authority );
-                                printFromClient
-                                    .OrderBy( l => l.PersonId )
-                                    .ThenBy( l => l.Order )
-                                    .ToList()
-                                    .ForEach( l => l.LabelFile = urlRoot + l.LabelFile );
-                                AddLabelScript( printFromClient.ToJson() );
-                            }
-
-                            if ( printFromServer.Any() )
-                            {
-                                var messages = ZebraPrint.PrintLabels( printFromServer );
-                            }
-
-                        }
-                        catch ( Exception ex )
-                        {
-                            LogException( ex );
-                        }
-                    }
+                    var queryParams = new Dictionary<string, string>();
+                    queryParams.Add( "PersonId", person.Id.ToString() );
+                    NavigateToCurrentPage( queryParams );
                 }
             }
-
-            NavigateToHomePage();
         }
 
         /// <summary>
@@ -378,5 +199,10 @@ namespace RockWeb.Plugins.com_bemaservices.CheckIn
         }
 
         #endregion
+
+        protected void btnReturn_Click( object sender, EventArgs e )
+        {
+            NavigateToHomePage();
+        }
     }
 }
